Move AvatarView to separate file
